% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everypoliticianr_core.R
\name{approx_date}
\alias{approx_date}
\title{Approximate Date fixer function}
\usage{
approx_date(date_string)
}
\arguments{
\item{date_string}{#popolo object to get lookup from
approx_date()}
}
\description{
Not all everypolitician dates are complete,
This coerces incomplete dates to the first day in month or year.
}
