% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everypoliticianr_core.R
\name{ep_alt_name_list}
\alias{ep_alt_name_list}
\title{Alternative Names}
\usage{
ep_alt_name_list(popolo)
}
\arguments{
\item{popolo}{#popolo object to get lookup from}
}
\description{
Expands all alternative names + normal names to create data.frame that links names to ep_ids
}
