% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everypoliticianr_core.R
\name{ep_get_countries_list}
\alias{ep_get_countries_list}
\title{Get countries list}
\usage{
ep_get_countries_list(json_location = default_json)
}
\arguments{
\item{json_location}{#optional - location to look for countries.json
ep_get_countries_list()}
}
\description{
Returns and caches the countries data frame
}
