% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/everypoliticianr_core.R
\name{ep_id_lookup}
\alias{ep_id_lookup}
\title{EveryPolitician ID Lookup}
\usage{
ep_id_lookup(popolo, scheme)
}
\arguments{
\item{popolo}{#popolo object to get lookup from}

\item{scheme}{alternate ID scheme. e.g use 'parlparse' to create a two-column lookup between id and the parlparse ids.
ep_id_lookup()}
}
\description{
Creates a lookup data frame from the alternate identifiers stored in the popolo file.
}
